requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
torch>=2.6.0
torchvision>=0.23.0
pandas>=2.0.3
numpy>=1.24.3
plotly>=5.17.0
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
python-socketio>=5.9.0
python-engineio>=4.7.1
APScheduler>=3.10.4
python-dotenv>=1.0.0
bcrypt>=4.0.1
requests>=2.31.0
